д.з. за 7-ой урок
