+++ app/config/routing.php
@@ -1,8 +1,10 @@
 <?php
 
 use Controller\MainController;
-use Controller\OrderController;
-use Controller\ProductController;
+use Controller\OrderInfoController;
+use Controller\OrderCheckoutController;
+use Controller\ProductInfoController;
+use Controller\ProductListController;
 use Symfony\Component\Routing\RouteCollection;
 use Symfony\Component\Routing\Route;
 
@@ -15,29 +17,29 @@ $routes->add(
 
 $routes->add(
     'product_list',
-    new Route('/product/list', ['_controller' => [ProductController::class, 'listAction']])
+    new Route('/product/list', ['_controller' => [ProductListController::class, 'listAction']])
 );
 $routes->add(
     'product_info',
-    new Route('/product/info/{id}', ['_controller' => [ProductController::class, 'infoAction']])
+    new Route('/product/info/{id}', ['_controller' => [ProductInfoController::class, 'infoAction']])
 );
 
 $routes->add(
     'order_info',
-    new Route('/order/info', ['_controller' => [OrderController::class, 'infoAction']])
+    new Route('/order/info', ['_controller' => [OrderInfoController::class, 'infoAction']])
 );
 $routes->add(
     'order_checkout',
-    new Route('/order/checkout', ['_controller' => [OrderController::class, 'checkoutAction']])
+    new Route('/order/checkout', ['_controller' => [OrderCheckoutController::class, 'checkoutAction']])
 );
 
 $routes->add(
     'user_authentication',
-    new Route('/user/authentication', ['_controller' => [\Controller\UserController::class, 'authenticationAction']])
+    new Route('/user/authentication', ['_controller' => [\Controller\UserAuthenticationController::class, 'authenticationAction']])
 );
 $routes->add(
     'logout',
-    new Route('/user/logout', ['_controller' => [\Controller\UserController::class, 'logoutAction']])
+    new Route('/user/logout', ['_controller' => [\Controller\UserLogoutController::class, 'logoutAction']])
 );
 
 return $routes;
