+++ src/Controller/OrderInfoController.php
@@ -11,7 +11,7 @@ use Service\User\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
-class OrderController
+class OrderInfoController
 {
     use Render;
 
@@ -33,21 +33,4 @@ class OrderController
         return $this->render('order/info.html.php', ['productList' => $productList, 'isLogged' => $isLogged]);
     }
 
-    /**
-     * Оформление заказа
-     *
-     * @param Request $request
-     * @return Response
-     */
-    public function checkoutAction(Request $request): Response
-    {
-        $isLogged = (new Security($request->getSession()))->isLogged();
-        if (!$isLogged) {
-            return $this->redirect('user_authentication');
-        }
-
-        (new BasketFacade($request->getSession()))->checkout();
-
-        return $this->render('order/checkout.html.php');
-    }
 }
