+++ src/Controller/ProductInfoController.php
@@ -11,7 +11,7 @@ use Service\Product\Product;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
-class ProductController
+class ProductInfoController
 {
     use Render;
 
@@ -41,17 +41,5 @@ class ProductController
         return $this->render('product/info.html.php', ['productInfo' => $productInfo, 'isInBasket' => $isInBasket]);
     }
 
-    /**
-     * Список всех продуктов
-     *
-     * @param Request $request
-     *
-     * @return Response
-     */
-    public function listAction(Request $request): Response
-    {
-        $productList = (new Product())->getAll($request->query->get('sort', ''));
 
-        return $this->render('product/list.html.php', ['productList' => $productList]);
-    }
 }
