+++ src/Controller/UserAuthenticationController.php
@@ -9,7 +9,7 @@ use Service\User\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
-class UserController
+class UserAuthenticationController
 {
     use Render;
 
@@ -39,16 +39,4 @@ class UserController
         return $this->render('user/authentication.html.php', ['error' => $error ?? '']);
     }
 
-    /**
-     * Выходим из системы
-     *
-     * @param Request $request
-     * @return Response
-     */
-    public function logoutAction(Request $request): Response
-    {
-        (new Security($request->getSession()))->logout();
-
-        return $this->redirect('index');
-    }
 }
